refactor(provider): extract AnyClient/AnyModel dispatch into dispatch.rs (dirge-4y4l stage 6)
Move the rig client/model dispatch enums out of provider/mod.rs into a
new provider::dispatch submodule: AnyClient (completion_model,
compress_messages), AnyModel (btw_query, build_stream_fn, name), and
the cfg(test) filter_tool_names helper. mod.rs re-exports via `pub use
dispatch::*` so provider::AnyClient / provider::AnyModel paths are
unchanged.

Pruned the now-unused mod.rs imports (CompletionClient, Prompt,
agent::prompt, SessionMessage); mod_tests.rs imports CompletionClient
directly (previously inherited via use super::*). Updated the
btw_query source-inspection test in task.rs to point at dispatch.rs.

Behavior-preserving. mod.rs 1738->1467. Builds clean -D warnings on
default + windows-default; 2374 tests green.
